+++ CHANGES.rst
@@ -26,6 +26,11 @@ Changelog
   [gbastien]
 - Display potential session title in prepare for signature forms.
   [gbastien]
+- In `@@store-items-template-as-annex-batch-action` display error message if
+  a template could not be generated because of technical error
+  (error durint generation) or condition not available for an item instead
+  breaking without any message and reloading the page.
+  [gbastien]
 
 4.2.29rc9 (2026-06-02)
 ----------------------